Reorganize some variables
