--- openfisca_us/variables/gov/irs/taxcalc/deductions/exemptions.py
@@ -1,49 +1,0 @@
-from openfisca_us.model_api import *
-
-
-class pre_c04600(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "Personal exemption before phase-out"
-    reference = "https://www.law.cornell.edu/uscode/text/26/151"
-    unit = USD
-
-    def formula(tax_unit, period, parameters):
-        exemption = parameters(period).irs.income.exemption
-        return where(
-            tax_unit("dsi", period),
-            0,
-            tax_unit("xtot", period) * exemption.amount,
-        )
-
-
-class exemption_phase_out_start(Variable):
-    value_type = float
-    entity = TaxUnit
-    label = "Exemption phase-out start"
-    definition_period = YEAR
-    unit = USD
-
-    def formula(tax_unit, period, parameters):
-        return parameters(period).irs.income.exemption.phase_out.start[
-            tax_unit("filing_status", period)
-        ]
-
-
-class c04600(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    documentation = "Personal exemptions after phase-out"
-    unit = USD
-
-    def formula(tax_unit, period, parameters):
-        phase_out = parameters(period).irs.income.exemption.phase_out
-        phase_out_start = tax_unit("exemption_phase_out_start", period)
-        line_5 = max_(
-            0, tax_unit("adjusted_gross_income", period) - phase_out_start
-        )
-        line_6 = line_5 / (2500 / tax_unit("sep", period))
-        line_7 = phase_out.rate * line_6
-        return tax_unit("pre_c04600", period) * (1 - line_7)
