--- openfisca_us/variables/gov/irs/taxcalc/deductions/qbided.py
@@ -1,1 +1,0 @@
-from openfisca_us.model_api import *
