+++ openfisca_us/variables/gov/irs/taxcalc/outputs.py
@@ -7,10 +7,9 @@ class sey(Variable):
     definition_period = YEAR
     unit = USD
 
-    def formula(person, period, parameters):
-        return add(
-            person, period, ["self_employment_income", "farm_income", "k1bx14"]
-        )
+    formula = sum_of_variables(
+        ["self_employment_income", "farm_income", "k1bx14"]
+    )
 
 
 class filer_sey(Variable):
